Use go:build to avoid using windows package on non-Windows plats
